Move Frontier into UpdateCheck
