+++ src/XIVLauncher.Core/Configuration/ILauncherConfig.cs
@@ -12,8 +12,6 @@ public interface ILauncherConfig
 
     public bool? DoVersionCheck { get; set; }
 
-    public bool? DoFrontierCheck { get; set; }
-
     public float? FontPxSize { get; set; }
 
     public string? CurrentAccountId { get; set; }
