+++ src/XIVLauncher.Core/XIVLauncher.Core.csproj
@@ -9,7 +9,7 @@
     <LangVersion>latest</LangVersion>
     <ApplicationIcon>Resources\dalamud_icon.ico</ApplicationIcon>
 
-    <Version>1.0.6.0</Version>
+    <Version>1.0.5.0</Version>
     <FileVersion>$(Version)</FileVersion>
     <AssemblyVersion>$(Version)</AssemblyVersion>
